services/django/pyproject.toml
[project]
name = "tryb3-backend"
version = "0.1.0"
description = "Tryb3 Backend API - Django REST + GraphQL + WebSockets"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "strawberry-graphql>=0.284.2",
    "uvicorn>=0.38.0",
]
